Introduce grid engine autoscaler library
