+++ workflows/pipe-common/scripts/generate_sge_profiles.py
@@ -14,12 +14,13 @@
 
 import logging
 import os
+
 import sys
 
 from pipeline.api import PipelineAPI
+from pipeline.hpc.autoscaler import GridEngineParameters
 from pipeline.log.logger import LocalLogger, RunLogger, TaskLogger, LevelLogger
 from pipeline.utils.profile import suffix_non_unique, build_environment_profiles
-from scripts.autoscale_sge import GridEngineParameters
 
 PROFILE_QUEUE_FORMAT = 'sge_profile_{}_queue.sh'
 PROFILE_QUEUE_PATTERN = '^sge_profile_(.+)_queue\\.sh$'
