+++ workflows/pipe-common/test/test_scale_down_handler.py
@@ -17,7 +17,8 @@ import logging
 from datetime import datetime
 from mock import MagicMock, Mock
 
-from scripts.autoscale_sge import GridEngineScaleDownHandler, GridEngineJob, GridEngineJobState, ComputeResource
+from pipeline.hpc.autoscaler import GridEngineScaleDownHandler, GridEngineJob, GridEngineJobState, \
+    ComputeResource
 from utils import assert_first_argument_contained, assert_first_argument_not_contained
 
 logging.basicConfig(level=logging.DEBUG, format='%(asctime)s [%(threadName)s] [%(levelname)s] %(message)s')
