+++ workflows/pipe-common/test/test_scale_up.py
@@ -17,8 +17,8 @@ import logging
 from datetime import datetime, timedelta
 from mock import MagicMock, Mock
 
-from scripts.autoscale_sge import GridEngineAutoscaler, GridEngineJob, GridEngineJobState, Clock, MemoryHostStorage, \
-    IntegralDemand
+from pipeline.hpc.autoscaler import GridEngineAutoscaler, GridEngineJob, GridEngineJobState, Clock, \
+    MemoryHostStorage, IntegralDemand
 
 logging.basicConfig(level=logging.DEBUG, format='%(asctime)s [%(threadName)s] [%(levelname)s] %(message)s')
