+++ workflows/pipe-common/test/test_worker_recorder.py
@@ -17,7 +17,7 @@ from datetime import datetime
 
 from mock import MagicMock, Mock
 
-from scripts.autoscale_sge import CloudPipelineWorkerRecorder, AvailableInstanceEvent, \
+from pipeline.hpc.autoscaler import CloudPipelineWorkerRecorder, AvailableInstanceEvent, \
     InsufficientInstanceEvent, FailingInstanceEvent
 
 logging.basicConfig(level=logging.DEBUG, format='%(asctime)s [%(threadName)s] [%(levelname)s] %(message)s')
